+++ packages/atomic/src/components/common/facets/category-facet/value-link.tsx
@@ -1,7 +1,7 @@
 import {FunctionalComponent, h} from '@stencil/core';
 import {i18n} from 'i18next';
-import {FacetValueLabelHighlight} from '../facet-value-label-highlight/facet-value-label-highlight';
-import {FacetValueLink} from '../facet-value-link/facet-value-link';
+import {FacetValueLabelHighlight} from '../facet-value-label-highlight/stencil-facet-value-label-highlight';
+import {FacetValueLink} from '../facet-value-link/stencil-facet-value-link';
 
 export interface CategoryFacetValueLinkProps {
   displayValue: string;