+++ packages/atomic/src/components/insight/result-templates/atomic-insight-result-children/atomic-insight-result-children.tsx
@@ -10,7 +10,7 @@ import {
   InitializableComponent,
   InitializeBindings,
 } from '../../../../utils/initialization-utils';
-import {elementHasAncestorTag} from '../../../../utils/stencil-utils';
+import {elementHasAncestorTag} from '../../../../utils/utils';
 import {extractUnfoldedItem} from '../../../common/interface/item';
 import {
   ChildTemplatesContext,